pyproject.toml
[tool.poetry]
name = "nonebot-plugin-help"
version = "0.6.1"
description = "A general help lister for nonebot2 plugins"
authors = ["XZhouQD <X.Zhou.QD@hotmail.com>"]
license = "AGPL v3"
readme = "README.md"
homepage = "https://github.com/XZhouQD/nonebot-plugin-help"
repository = "https://github.com/XZhouQD/nonebot-plugin-help"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.8.0"
nonebot2 = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
